Salli duplikaattitallennus YTR-opiskeluoikeuksille;
YTR-downloadissa on sallittua päivittää olemassaolevaa ilman oidia, vaikkei Koskessa normaalisti olekaan. YTR downloadille on jo
oma luokkansa, joten siirrä tämä logiikka myös erikseen YTR-luokkaan.
